# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Timekeeping and Helpdesk API
  version: 1.0.0
paths:
...
  /api/v1/tasks/:
    get:
      operationId: tasks_list
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupedTaskList'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
    post:
      operationId: tasks_create
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              properties:
                assigned_date:
                  format: date
                  type: string
                assignee_ids:
                  items:
                    minimum: 1
                    type: integer
                  minItems: 1
                  type: array
                description:
                  default: ''
                  type: string
                expected_location:
                  default: ''
                  maxLength: 500
                  type: string
                location_id:
                  minimum: 1
                  type:
                  - integer
                  - 'null'
                title:
                  minLength: 1
                  type: string
              required:
              - title
              - assigned_date
              type: object
          application/x-www-form-urlencoded:
            schema:
              additionalProperties: false
              properties:
                assigned_date:
                  format: date
                  type: string
                assignee_ids:
                  items:
                    minimum: 1
                    type: integer
                  minItems: 1
                  type: array
                description:
                  default: ''
                  type: string
                expected_location:
                  default: ''
                  maxLength: 500
                  type: string
                location_id:
                  minimum: 1
                  type:
                  - integer
                  - 'null'
                title:
                  minLength: 1
                  type: string
              required:
              - title
              - assigned_date
              type: object
          multipart/form-data:
            schema:
              additionalProperties: false
              properties:
                assigned_date:
                  format: date
                  type: string
                assignee_ids:
                  items:
                    minimum: 1
                    type: integer
                  minItems: 1
                  type: array
                description:
                  default: ''
                  type: string
                expected_location:
                  default: ''
                  maxLength: 500
                  type: string
                location_id:
                  minimum: 1
                  type:
                  - integer
                  - 'null'
                title:
                  minLength: 1
                  type: string
              required:
              - title
              - assigned_date
              type: object
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InactiveAssigneeError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
  /api/v1/tasks/{task_id}/:
    delete:
      operationId: tasks_destroy
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          type: string
      responses:
        '204':
          description: No response body
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
    get:
      operationId: tasks_retrieve
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
    patch:
      operationId: tasks_partial_update
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InactiveAssigneeError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
  /api/v1/tasks/{task_id}/complete-field:
    post:
      operationId: tasks_complete_field_create
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          type: string
      - in: path
        name: task_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskFieldCompletion'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TaskFieldCompletion'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TaskFieldCompletion'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
  /api/v1/tasks/{task_id}/complete-override:
    post:
      operationId: tasks_complete_override_create
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskOverride'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TaskOverride'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TaskOverride'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskAlreadyCompletedError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
  /api/v1/tasks/{task_id}/evidence-uploads:
    post:
      operationId: tasks_evidence_uploads_create
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvidenceUpload'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/EvidenceUpload'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EvidenceUpload'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceUploadIntent'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
  /api/v1/tasks/{task_id}/photos/{photo_id}/access:
    post:
      operationId: tasks_photos_access_create
      parameters:
      - in: path
        name: photo_id
        required: true
        schema:
          type: string
      - in: path
        name: task_id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoAccess'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
  /api/v1/tasks/{task_id}/status:
    post:
      operationId: tasks_status_create
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskStatus'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TaskStatus'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TaskStatus'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
...
components:
  schemas:
...
    GroupedTaskList:
      properties:
        business_date:
          format: date
          type: string
        completed:
          items:
            $ref: '#/components/schemas/TaskItem'
          type: array
        overdue:
          items:
            $ref: '#/components/schemas/TaskItem'
          type: array
        today:
          items:
            $ref: '#/components/schemas/TaskItem'
          type: array
        upcoming:
          items:
            $ref: '#/components/schemas/TaskItem'
          type: array
      required:
      - business_date
      - completed
      - overdue
      - today
      - upcoming
      type: object
...
    InactiveAssigneeError:
      properties:
        details:
          $ref: '#/components/schemas/InactiveAssigneeDetails'
        error:
          type: string
        error_code:
          $ref: '#/components/schemas/InactiveAssigneeErrorErrorCodeEnum'
        message:
          type: string
        request_id:
          format: uuid
          type: string
      required:
      - details
      - error
      - error_code
      - message
      - request_id
      type: object
...
    TaskDetail:
      properties:
        assigned_date:
          format: date
          type: string
        assignees:
          items:
            $ref: '#/components/schemas/TaskAssignee'
          type: array
        block_reason:
          nullable: true
          type: string
        completed_at:
          format: date-time
          nullable: true
          type: string
        completed_by:
          allOf:
          - $ref: '#/components/schemas/TaskUser'
          nullable: true
        completion_method:
          nullable: true
          type: string
        completion_note:
          nullable: true
          type: string
        created_by:
          $ref: '#/components/schemas/TaskUser'
        description:
          type: string
        expected_location:
          type: string
        group:
          type: string
        id:
          type: integer
        location:
          allOf:
          - $ref: '#/components/schemas/TaskLocation'
          nullable: true
        overdue_days:
          nullable: true
          type: integer
        status:
          $ref: '#/components/schemas/TaskStatusEnum'
        title:
          type: string
        updates:
          items:
            $ref: '#/components/schemas/TaskLifecycleUpdate'
          type: array
      required:
      - assigned_date
      - assignees
      - block_reason
      - completed_at
      - completed_by
      - completion_method
      - completion_note
      - created_by
      - description
      - expected_location
      - group
      - id
      - location
      - overdue_days
      - status
      - title
      - updates
      type: object
    TaskError:
      properties:
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        error_code:
          type: string
        message:
          type: string
        request_id:
          format: uuid
          type: string
      required:
      - details
      - error
      - error_code
      - message
      - request_id
      type: object
    TaskFieldCompletion:
      properties:
        accuracy_m:
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
          type: string
        captured_at:
          format: date-time
          type: string
        completion_note:
          nullable: true
          type: string
        latitude:
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,15})?$
          type: string
        longitude:
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,15})?$
          type: string
        selected_location_id:
          minimum: 1
          nullable: true
          type: integer
        upload_ids:
          items:
            format: uuid
            type: string
          maxItems: 5
          minItems: 1
          type: array
      required:
      - accuracy_m
      - captured_at
      - latitude
      - longitude
      - upload_ids
      type: object
...
    TaskOverride:
      properties:
        completion_note:
          type: string
      required:
      - completion_note
      type: object
...
    TaskUpdate:
      properties:
        assignee_ids:
          items:
            minimum: 1
            type: integer
          type: array
        description:
          type: string
        expected_location:
          maxLength: 500
          type: string
        location_id:
          minimum: 1
          nullable: true
          type: integer
        title:
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http